(PC-16475) feat(tests): update tests and snapshots
