+++ src/libs/notifications/components/PushNotificationModal.native.test.tsx
@@ -3,12 +3,24 @@ import React from 'react'
 import { analytics } from 'libs/firebase/analytics'
 import { fireEvent, render } from 'tests/utils'
 
-import { PushNotificationsModal } from '../PushNotificationsModal'
+import { PushNotificationsModal } from './PushNotificationsModal'
+
+const onDismiss = jest.fn()
+const onRequestPermission = jest.fn()
 
 describe('PushNotificationsModal', () => {
+  it('should render properly', () => {
+    const renderAPI = render(
+      <PushNotificationsModal
+        onDismiss={onDismiss}
+        visible={true}
+        onRequestPermission={onRequestPermission}
+      />
+    )
+    expect(renderAPI).toMatchSnapshot()
+  })
+
   it('should open settings and log event logOpenNotificationSettings', () => {
-    const onDismiss = jest.fn()
-    const onRequestPermission = jest.fn()
     const renderAPI = render(
       <PushNotificationsModal
         onDismiss={onDismiss}
